Data Analysis with Songs and Lyrics/requirements.txt
lyricsgenius
statsmodel
numpy
scipy
bs4
billboard
pandas
matplotlib
sklearn
nlkt
wordcloud
spotipy
